Do the Intermediate Tutorial
I am having all sorts of trouble again with the database. It says
something about there not being a user_id column. When I try to refresh
the database, it says it can't find the CreateTasksTable class. I don't
know what that really means, but I don't know how to fix it.